Modify function prototype match to work with UEFI
UEFI aliases "VOID" in all caps to "void"; modify the script to recognize
this usage.
UEFI style function prototypes make heavy use of "IN" and "OUT" keywords
to indicate the use of function parameters. Add to the regex matching
groups so that the keywords are recognized and use it to generate similar
statements in the doxygen format.